+++ test/redis/ick_test.rb
@@ -1169,5 +1169,50 @@ class Redis
         end
       end
     end
+
+    def test_big_ick_performance
+      chunk_size   = 50000
+      num_chunks   = 3
+      s_and_ms     = (chunk_size*num_chunks).times.map{|i|[i,i.to_s]}.flatten
+      puts ""
+      puts "chunk_size:       #{chunk_size}"
+      puts "num_chunks:       #{num_chunks}"
+      puts "s_and_ms.size:    #{s_and_ms.size}"
+      puts "-" * 40
+      s_and_ms.each_slice(2*chunk_size) do |slice|
+        t0 = Time.now
+        ick.ickadd(@ick_key,*slice)
+        t1 = Time.now
+        puts "ickadd     t1-t0: %5.3f" % (t1-t0)
+      end
+      puts "-" * 40
+      num_chunks.times do
+        t0 = Time.now
+        ick.ickstats(@ick_key)
+        t1 = Time.now
+        puts "ickstats   t1-t0: %5.3f" % (t1-t0)
+      end
+      [false,true].each do |backwash|
+        puts "-" * 40
+        num_chunks.times do
+          t0 = Time.now
+          ick.ickreserve(@ick_key,chunk_size,backwash: backwash)
+          t1 = Time.now
+          puts "ickreserve t1-t0: %5.3f backwash: #{backwash}" % (t1-t0)
+        end
+      end
+      puts "-" * 40
+      num_chunks.times do
+        t0 = Time.now
+        msgs = ick.ickreserve(@ick_key,chunk_size,backwash: false).map(&:first)
+        t1 = Time.now
+        puts "ickreserve t1-t0: %5.3f backwash: false" % (t1-t0)
+        t0 = Time.now
+        num  = ick.ickcommit(@ick_key,*msgs)
+        t1 = Time.now
+        puts "ickcommit  t1-t0: %5.3f num: #{num}" % (t1-t0)
+      end
+    end
+
   end
 end